Create ImageDetails as part of the public API
