+++ jib-core/src/main/java/com/google/cloud/tools/jib/api/ImageDetails.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2022 Google LLC.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package com.google.cloud.tools.jib.api;
+
+import java.security.DigestException;
+import java.util.List;
+
+public interface ImageDetails {
+
+  long getSize();
+
+  DescriptorDigest getImageId() throws DigestException;
+
+  List<DescriptorDigest> getDiffIds() throws DigestException;
+}
