+++ jib-core/src/test/java/com/google/cloud/tools/jib/docker/CliDockerClientTest.java
@@ -18,8 +18,8 @@ package com.google.cloud.tools.jib.docker;
 
 import com.google.cloud.tools.jib.api.DescriptorDigest;
 import com.google.cloud.tools.jib.api.DockerClient;
-import com.google.cloud.tools.jib.api.DockerClient.DockerImageDetails;
 import com.google.cloud.tools.jib.api.ImageReference;
+import com.google.cloud.tools.jib.docker.CliDockerClient.DockerImageDetails;
 import com.google.cloud.tools.jib.image.ImageTarball;
 import com.google.cloud.tools.jib.json.JsonTemplateMapper;
 import com.google.common.collect.ImmutableMap;
